+++ src/language_gdnative.c
@@ -39,16 +39,21 @@ extern const char LUA_INIT_SCRIPT[];
 extern const size_t LUA_INIT_SCRIPT_SIZE;
 
 #define PLUGINSCRIPT_CALLBACKS_KEY "lps_callbacks"
-#define lps_push_callback(L, name) \
-	lua_getfield(L, LUA_REGISTRYINDEX, PLUGINSCRIPT_CALLBACKS_KEY), \
-	lua_getfield(L, -1, name), \
+#define LPS_PUSH_CALLBACK(L, name) \
+	lua_getfield(L, LUA_REGISTRYINDEX, PLUGINSCRIPT_CALLBACKS_KEY); \
+	lua_getfield(L, -1, name); \
 	lua_remove(L, -2)
+
+#define LPS_PCALL_CONSUME_ERROR(L, prefix) \
+	HGDN_PRINT_ERROR(prefix ": %s", lua_tostring(L, -1)); \
+	lua_pop(L, 1)
 
 #ifdef LUAJIT_DYNAMICALLY_LINKED
 // Active shared library path, for loading symbols in FFI
 static hgdn_string lps_active_library_path;
 #endif
-static bool in_editor;
+static bool lps_in_editor;
+static int lps_pcall_error_handler_index;
 static lua_State *lps_L;
 
 // Language functions
@@ -99,12 +104,10 @@ static int lps_lua_string_replace(lua_State *L) {
 	return 1;
 }
 
-#ifdef DEBUG
 static int lps_pcall_error_handler(lua_State *L) {
 	luaL_traceback(L, L, lua_tostring(L, 1), 0);
 	return 1;
 }
-#endif
 
 static godot_pluginscript_language_data *lps_language_init() {
 	lua_State *L = lps_L = lua_newstate(&lps_alloc, NULL);
@@ -118,36 +121,29 @@ static godot_pluginscript_language_data *lps_language_init() {
 	lua_setfield(L, -2, "replace");
 	lua_pop(L, 1);  // pop string
 
+	// maintain C pcall error handler in the stack, used by all callbacks
+	lua_pushcfunction(L, &lps_pcall_error_handler);
+	lps_pcall_error_handler_index = lua_gettop(L);
+
 	if (luaL_loadbufferx(L, LUA_INIT_SCRIPT, LUA_INIT_SCRIPT_SIZE - 1, "lps_init_script", "t") != LUA_OK) {
-		const char *error_msg = lua_tostring(L, -1);
-		HGDN_PRINT_ERROR("Error loading initialization script: %s", error_msg);
+		LPS_PCALL_CONSUME_ERROR(L, "Error loading Lua language initialization script");
 		return L;
 	}
-#ifdef DEBUG
-	lua_pushcfunction(L, &lps_pcall_error_handler);
-	const int err_handler_index = lua_gettop(L);
-	lua_pushvalue(L, -2);
-#else
-	const int err_handler_index = 0;
-#endif
+
 	// registry.lps_callbacks = {} // maintaining it on stack
 	lua_createtable(L, 0, 10);
 	lua_pushvalue(L, -1);
 	lua_setfield(L, LUA_REGISTRYINDEX, PLUGINSCRIPT_CALLBACKS_KEY);
 
-	lua_pushboolean(L, in_editor);
+	lua_pushboolean(L, lps_in_editor);
 #ifdef LUAJIT_DYNAMICALLY_LINKED
 	lua_pushlstring(L, lps_active_library_path.ptr, lps_active_library_path.length);
 #else
 	lua_pushnil(L);
 #endif
-	if (lua_pcall(L, 3, 0, err_handler_index) != LUA_OK) {
-		const char *error_msg = lua_tostring(L, -1);
-		HGDN_PRINT_ERROR("Error running initialization script: %s", error_msg);
+	if (lua_pcall(L, 3, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(L, "Error in Lua language initialization script");
 	}
-#ifdef DEBUG
-	lua_pop(L, 2);
-#endif
 	return L;
 }
 
@@ -157,10 +153,12 @@ static void lps_language_finish(godot_pluginscript_language_data *data) {
 }
 
 static void lps_language_add_global_constant(godot_pluginscript_language_data *data, const godot_string *name, const godot_variant *value) {
-	lps_push_callback(lps_L, "language_add_global_constant");
+	LPS_PUSH_CALLBACK(lps_L, "language_add_global_constant");
 	lua_pushlightuserdata(lps_L, (void *) name);
 	lua_pushlightuserdata(lps_L, (void *) value);
-	lua_pcall(lps_L, 2, 0, 0);
+	if (lua_pcall(lps_L, 2, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in add_global_constant");
+	}
 }
 
 // Script manifest
@@ -177,12 +175,14 @@ static godot_pluginscript_script_manifest lps_script_init(godot_pluginscript_lan
 	hgdn_core_api->godot_array_new(&manifest.properties);
 
 	godot_error cb_error = GODOT_ERR_SCRIPT_FAILED;
-	lps_push_callback(lps_L, "script_init");
+	LPS_PUSH_CALLBACK(lps_L, "script_init");
 	lua_pushlightuserdata(lps_L, (void *) &manifest);
 	lua_pushlightuserdata(lps_L, (void *) path);
 	lua_pushlightuserdata(lps_L, (void *) source);
 	lua_pushlightuserdata(lps_L, (void *) &cb_error);
-	lua_pcall(lps_L, 4, 0, 0);
+	if (lua_pcall(lps_L, 4, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in script_init");
+	}
 	if (error) {
 		*error = cb_error;
 	}
@@ -191,35 +191,42 @@ static godot_pluginscript_script_manifest lps_script_init(godot_pluginscript_lan
 }
 
 static void lps_script_finish(godot_pluginscript_script_data *data) {
-	lps_push_callback(lps_L, "script_finish");
+	LPS_PUSH_CALLBACK(lps_L, "script_finish");
 	lua_pushlightuserdata(lps_L, (void *) data);
-	lua_pcall(lps_L, 1, 0, 0);
+	if (lua_pcall(lps_L, 1, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in script_finish");
+	}
 }
 
 
 // Instance
 static godot_pluginscript_instance_data *lps_instance_init(godot_pluginscript_script_data *data, godot_object *owner) {
-	lps_push_callback(lps_L, "instance_init");
+	LPS_PUSH_CALLBACK(lps_L, "instance_init");
 	lua_pushlightuserdata(lps_L, (void *) data);
 	lua_pushlightuserdata(lps_L, (void *) owner);
-	if (lua_pcall(lps_L, 2, 0, 0) != LUA_OK) {
-		return NULL;
+	void *result = NULL;
+	lua_pushlightuserdata(lps_L, (void *) &result);
+	if (lua_pcall(lps_L, 3, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_init");
 	}
-	return owner;
+	return result;
 }
 
 static void lps_instance_finish(godot_pluginscript_instance_data *data) {
-	lps_push_callback(lps_L, "instance_finish");
+	LPS_PUSH_CALLBACK(lps_L, "instance_finish");
 	lua_pushlightuserdata(lps_L, (void *) data);
-	lua_pcall(lps_L, 1, 0, 0);
+	if (lua_pcall(lps_L, 1, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_finish");
+	}
 }
 
 static godot_bool lps_instance_set_prop(godot_pluginscript_instance_data *data, const godot_string *name, const godot_variant *value) {
-	lps_push_callback(lps_L, "instance_set_prop");
+	LPS_PUSH_CALLBACK(lps_L, "instance_set_prop");
 	lua_pushlightuserdata(lps_L, (void *) data);
 	lua_pushlightuserdata(lps_L, (void *) name);
 	lua_pushlightuserdata(lps_L, (void *) value);
-	if (lua_pcall(lps_L, 3, 1, 0) != LUA_OK) {
+	if (lua_pcall(lps_L, 3, 1, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_set_prop");
 		return false;
 	}
 	godot_bool have_prop = lua_toboolean(lps_L, -1);
@@ -228,11 +235,12 @@ static godot_bool lps_instance_set_prop(godot_pluginscript_instance_data *data, 
 }
 
 static godot_bool lps_instance_get_prop(godot_pluginscript_instance_data *data, const godot_string *name, godot_variant *value_ret) {
-	lps_push_callback(lps_L, "instance_get_prop");
+	LPS_PUSH_CALLBACK(lps_L, "instance_get_prop");
 	lua_pushlightuserdata(lps_L, (void *) data);
 	lua_pushlightuserdata(lps_L, (void *) name);
 	lua_pushlightuserdata(lps_L, (void *) value_ret);
-	if (lua_pcall(lps_L, 3, 1, 0) != LUA_OK) {
+	if (lua_pcall(lps_L, 3, 1, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_get_prop");
 		return false;
 	}
 	godot_bool have_prop = lua_toboolean(lps_L, -1);
@@ -245,45 +253,52 @@ static godot_variant lps_instance_call_method(godot_pluginscript_instance_data *
 	if (error) {
 		error->error = GODOT_CALL_ERROR_CALL_ERROR_INVALID_METHOD;
 	}
-	lps_push_callback(lps_L, "instance_call_method");
+	LPS_PUSH_CALLBACK(lps_L, "instance_call_method");
 	lua_pushlightuserdata(lps_L, (void *) data);
 	lua_pushlightuserdata(lps_L, (void *) method);
 	lua_pushlightuserdata(lps_L, (void *) args);
 	lua_pushinteger(lps_L, argcount);
 	lua_pushlightuserdata(lps_L, (void *) &var);
 	lua_pushlightuserdata(lps_L, (void *) error);
-	lua_pcall(lps_L, 6, 0, 0);
+	if (lua_pcall(lps_L, 6, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_call_method");
+	}
 	return var;
 }
 
 static void lps_instance_notification(godot_pluginscript_instance_data *data, int notification) {
-	lps_push_callback(lps_L, "instance_notification");
+	LPS_PUSH_CALLBACK(lps_L, "instance_notification");
 	lua_pushlightuserdata(lps_L, (void *) data);
 	lua_pushinteger(lps_L, notification);
-	lua_pcall(lps_L, 2, 0, 0);
+	if (lua_pcall(lps_L, 2, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in instance_notification");
+	}
 }
 
 // In-editor callbacks
 godot_string lps_get_template_source_code(godot_pluginscript_language_data *data, const godot_string *class_name, const godot_string *base_class_name) {
 	godot_string ret;
 	hgdn_core_api->godot_string_new(&ret);
-	lps_push_callback(lps_L, "get_template_source_code");
+	LPS_PUSH_CALLBACK(lps_L, "get_template_source_code");
 	lua_pushlightuserdata(lps_L, (void *) class_name);
 	lua_pushlightuserdata(lps_L, (void *) base_class_name);
 	lua_pushlightuserdata(lps_L, (void *) &ret);
-	lua_pcall(lps_L, 3, 0, 0);
+	if (lua_pcall(lps_L, 3, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in get_template_source_code");
+	}
 	return ret;
 }
 
 godot_bool lps_validate(godot_pluginscript_language_data *data, const godot_string *script, int *line_error, int *col_error, godot_string *test_error, const godot_string *path, godot_pool_string_array *functions) {
-	lps_push_callback(lps_L, "validate");
+	LPS_PUSH_CALLBACK(lps_L, "validate");
 	lua_pushlightuserdata(lps_L, (void *) script);
 	lua_pushlightuserdata(lps_L, (void *) line_error);
 	lua_pushlightuserdata(lps_L, (void *) col_error);
 	lua_pushlightuserdata(lps_L, (void *) test_error);
 	lua_pushlightuserdata(lps_L, (void *) path);
 	lua_pushlightuserdata(lps_L, (void *) functions);
-	if (lua_pcall(lps_L, 6, 1, 0) != LUA_OK) {
+	if (lua_pcall(lps_L, 6, 1, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in validate");
 		return false;
 	}
 	godot_bool success = lua_toboolean(lps_L, -1);
@@ -294,12 +309,14 @@ godot_bool lps_validate(godot_pluginscript_language_data *data, const godot_stri
 godot_string lps_make_function(godot_pluginscript_language_data *data, const godot_string *class_name, const godot_string *name, const godot_pool_string_array *args) {
 	godot_string ret;
 	hgdn_core_api->godot_string_new(&ret);
-	lps_push_callback(lps_L, "make_function");
+	LPS_PUSH_CALLBACK(lps_L, "make_function");
 	lua_pushlightuserdata(lps_L, (void *) class_name);
 	lua_pushlightuserdata(lps_L, (void *) name);
 	lua_pushlightuserdata(lps_L, (void *) args);
 	lua_pushlightuserdata(lps_L, (void *) &ret);
-	lua_pcall(lps_L, 4, 0, 0);
+	if (lua_pcall(lps_L, 4, 0, lps_pcall_error_handler_index) != LUA_OK) {
+		LPS_PCALL_CONSUME_ERROR(lps_L, "Error in make_function");
+	}
 	return ret;
 }
 
@@ -362,8 +379,8 @@ GDN_EXPORT void PREFIX_SYMBOL(gdnative_init)(godot_gdnative_init_options *option
 		return;
 	}
 
-	in_editor = options->in_editor;
-	if (in_editor) {
+	lps_in_editor = options->in_editor;
+	if (lps_in_editor) {
 		lps_register_in_editor_callbacks(&lps_language_desc);
 	}