+++ src/lua_string_extras.lua
@@ -24,15 +24,6 @@
 --- Extra functionality for Lua's `string` library.
 -- @module string_extras
 
-local function string_join(sep, ...)
-	local result = {}
-	for i = 1, select('#', ...) do
-		local s = select(i, ...)
-		table_insert(result, tostring(s))
-	end
-	return table_concat(result, sep)
-end
-
 --- Returns a Lua string with all values joined by `separator`.
 -- `tostring` is called to each of the passed values before joining.
 -- @usage
@@ -41,7 +32,24 @@ end
 -- @tparam string sep
 -- @param ...  Values to be joined, stringified by `tostring`.
 -- @treturn string
+local function string_join(sep, ...)
+	local result = {}
+	for i = 1, select('#', ...) do
+		local s = select(i, ...)
+		table_insert(result, tostring(s))
+	end
+	return table_concat(result, sep)
+end
 string.join = string_join
+
+--- Quote a value, alias for `string.format("%q", tostring(value))`.
+-- @function string.quote
+-- @param value
+-- @treturn string
+local function string_quote(value)
+	return string_format('%q', tostring(value))
+end
+string.quote = string_quote
 
 --- Performs plain substring substitution, with no characters in `pattern` or `replacement` being considered magic.
 -- @usage